pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project]
name = "threat-intelligence-api"
version = "0.1.0"
description = "FastAPI backend for a real-time security dashboard with CQRS architecture"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "opensearch-py[async]==2.8.0",
    "asyncpg==0.30.0",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.14.1",
    "redis[hiredis]==5.2.1",
    "structlog==24.4.0",
    "pydantic-settings==2.7.1",
    "psycopg2-binary==2.9.10",
]

[project.optional-dependencies]
dev = [
    "httpx==0.28.1",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.3",
    "ruff==0.9.7",
    "mypy==1.14.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]
extend-exclude = ["src/alembic/versions"]

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "E", "W", # pycodestyle
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]
ignore = [
    "B008",   # function calls in defaults (standard FastAPI pattern for Query/Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["src/alembic/"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["opensearchpy.*", "asyncpg.*"]
ignore_missing_imports = true
